misc: fixed a perfcct error when pc greater than 0x80000000.
pc is unsigned, but sqlite3 only supports signed integer [-2^63, 2^63-1], if real pc > 2^63-1, it will be stored as negative number, (negtive pc = real pc - 2^64), when read a negtive pc, real pc = negtive pc + 2^64.

+++ src/cpu/o3/perfCCT.cc
@@ -77,7 +77,11 @@ PerfCCT::commitMeta(InstSeqNum sn)
     }
     // dump string last
     ss << ",\'" << meta->disasm << "\'";
-    ss << "," << meta->pc;
+    // pc is unsigned, but sqlite3 only supports signed integer [-2^63, 2^63-1]
+    // if real pc > 2^63-1, it will be stored as negative number
+    // (negtive pc = real pc - 2^64)
+    // when read a negtive pc, real pc = negtive pc + 2^64
+    ss << "," << int64_t(meta->pc);
     ss << ");";
     archdb->execmd(ss.str());
     ss.str(std::string());
